Implemented Hook
rewrote code to use hooks  as suggested in #17
merged files into one
updated readme
